Add Python files about the merging and creation of data sets
Merging_data_sets.ipynb: Merges the csv data sets corresponding to fire, ndvi, soil moisture and weather conditions for Portugal with the tessellation. Creates a big data set for the period of time 01.01.2020 until 31.03.2025, and also single data sets with the data of the years 2020, 2021, 2022, 2023, and 2024. 

creation_data_set_fires_Montesinho_park.ipynb: Feature Engineering and Creation of Data Sets for Montesinho Park daily and weekly for the years 2020 to March 2025. It starts with the data set montesinho_clean_OriginalFeatures.csv that was previously cleaned and completed. 

list_features_dataset_Portugal.ipynb: This contains the list of all the features in the dataset allfeat_portugal.csv, and the number of missing values for each feature.